samgrabby/requirements/local.txt
-r common.txt

Werkzeug[watchdog]==3.0.1
ipdb==0.13.13

# Django
django-debug-toolbar==4.2.0
django-browser-reload==1.12.1

# Code quality
pylint-django==2.5.5
pre-commit==3.6.0
